can: Add basic types
This lays the groundwork for the upcoming high level CAN driver
implementation.
The types are directly modelled after STM32 register and bit layouts.
With `#[repr(C)]` the types could be copied directly to the memory mapped
CAN registers. It has to be investigated if this makes sense considering
the tradeoff between performance and the use of unsafe code.