Add support for reconfiguring a Serial port
Example use case is to implement a 1-Wire Bus "Master" [1]. In that
example, the UART is used instead of bit-banging the 1-Wire protocol.
However, not all signals can be generated with a single baud rate. In
particular, the 1-wire reset strobe is much longer than the 1-wire data
strobes, requiring to use a different, slower baud rate to reset
devices.

   [1]: https://www.maximintegrated.com/en/design/technical-documents/tutorials/2/214.html

Signed-off-by: Jonas Schäfer <j.wielicki@sotecware.net>