Support pin type erasure in Serial
There are many possible pin combinations for a Serial peripheral. Many
support at least one alternative pin group to be used, and in addition
to that, each pin can also be configured differently. For instance,
instead of the common Output<PushPull> with an Input<Floating>, one
could also use Output<OpenDrain> and Input<PullUp>, e.g. when
interfacing with a 1-Wire bus.

This pin information is already lost when splitting the Serial object
into Rx/Tx. In order to allow a reunification into a Serial object which
represents ownership of both Tx and Rx, we need a Serial variant which
has that pin information erased.