Add possibility to restore running rtc
So far the Rtc had to be reset on every
startup, leading to time lacking behind.

This has been addressed by adding method
`Rtc::restore_or_new` that allows to restore
the previous rtc, without resetting it.
In case the rtc was not running, it is initialized.

Documentation has been added and it was included
in `Rtc::new` that `Rtc::restore_or_new` should be
used for persistent time accuracy.

Example `examples/rtc.rs` has been updated to use
the new method.